+++ app/Http/Controllers/Backend/ProductController.php
@@ -6,10 +6,13 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StoreProductRequest;
 use App\Models\Product;
 use App\Models\Category;
 use App\Models\Image;
+use App\Models\Sale;
+use App\Models\User;
 
 
 class ProductController extends Controller
@@ -21,8 +24,15 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = Product::orderBy('updated_at', 'DESC')->get();
-        return view('backend.products.index', ['products' => $products]);
+        $user = Auth::user();
+        if ($user->can('viewAny', Product::class)) {
+            $products = Product::orderBy('updated_at', 'DESC')
+                                ->with(['sale', 'category', 'user'])
+                                ->get();
+            return view('backend.products.index', ['products' => $products]);
+        } else {
+            return redirect(route('frontend.index'));
+        }
     }
 
     /**
@@ -32,8 +42,14 @@ class ProductController extends Controller
      */
     public function create()
     {
-        $categories = Category::get();
-        return view('backend.products.create', ['categories' => $categories]);
+        $user = Auth::user();
+        if ($user->can('create', Product::class)){
+            $categories = Category::get();
+            return view('backend.products.create', ['categories' => $categories]);
+        }else {
+            return redirect(route('frontend.index'));
+        }
+        
     }
 
     /**
@@ -52,15 +68,10 @@ class ProductController extends Controller
         $file = $request->file('thumbnail');
         $path = Storage::disk('public')->putFileAs('thumbnail_product', $file, $file->getClientOriginalName());
         $product->thumbnail = $path;
-        // $product->origin_price = $request->get('origin_price');
-        // $product->sale_price = $request->get('sale_price');
-        // $product->discount_percent = $request->get('discount_percent');
         $product->description = $request->get('description');
         $product->status = $request->get('status');
         $product->user_id = Auth::user()->id;
         $product->save();
-
-        
 
         if ($request->hasFile('images')) {
             $images = $request->file('images');
@@ -73,6 +84,13 @@ class ProductController extends Controller
             }
         }
 
+        $sale = new Sale();
+        $sale->product_id = $product->id;
+        $sale->origin_price = $request->origin_price;
+        $sale->sale_price = $request->sale_price;
+        $sale->discount_percent = $request->discount_percent;
+        $sale->save();
+
         return redirect()->route('backend.product.index');
         
     }
@@ -83,14 +101,19 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Product $product)
     {
-        $product = Product::find($id);
-        $images = $product->images()->where('product_id', $id)->get();
-        return view('backend.products.show', [
-            'product' => $product,
-            'images' =>$images
-            ]);
+        $user = Auth::user();
+
+        if ($user->can('view', $product)){
+            $images = $product->images()->get();
+            return view('backend.products.show', [
+                'product'   => $product,
+                'images'    => $images
+                ]);
+        }else {
+            return redirect(route('frontend.index'));
+        }
     }
 
     /**
@@ -99,14 +122,21 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Product $product)
     {
-        $product = Product::find($id);
-        $categories = Category::get();
-        return view('backend.products.edit', [
-            'product' => $product,
-            'categories' => $categories
-            ]);
+        $user = Auth::user();
+
+        if ($user->can('update', $product)) {
+            $sale = $product->sale;
+            $categories = Category::get();
+            return view('backend.products.edit', [
+                'product'       => $product,
+                'categories'    => $categories,
+                'sale'          => $sale,
+                ]);
+        }else {
+            return redirect()->route('backend.product.index');
+        };
     }
 
     /**
@@ -123,9 +153,9 @@ class ProductController extends Controller
         $product->slug = \Illuminate\Support\Str::slug($request->get('name'));
         $product->quantity = $request->get('quantity');
         $product->category_id = $request->get('category_id');
-        $product->origin_price = $request->get('origin_price');
-        $product->sale_price = $request->get('sale_price');
-        $product->discount_percent = $request->get('discount_percent');
+        // $product->origin_price = $request->get('origin_price');
+        // $product->sale_price = $request->get('sale_price');
+        // $product->discount_percent = $request->get('discount_percent');
         $product->content = $request->get('content');
         $product->status = $request->get('status');
         $product->user_id = Auth::user()->id;
@@ -140,26 +170,32 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Product $product)
     {
-        try {
-            $product = Product::find($id);
-            $success = $product->delete();
-
-            if($success){
+        $user = Auth::user();
+        if ($user->can('delete', $product)) {
+            try {
+                $success = $product->delete();
+    
+                if($success){
+                    return response()->json([
+                        'error'=>false,
+                        'message'=>"Đã xóa",
+                    ]);
+                    location.reload();
+                }
+    
+            }catch (\Exception $e){
+                $message = "Xóa không thành công";
                 return response()->json([
-                    'error'=>false,
-                    'message'=>"Đã xóa",
+                    'error'=>true,
+                    'message'=>$e->getMessage(),
                 ]);
-                location.reload();
             }
-
-        }catch (\Exception $e){
-            $message = "Xóa không thành công";
-            return response()->json([
-                'error'=>true,
-                'message'=>$e->getMessage(),
-            ]);
+        } else {
+            return redirect()->route('backend.product.index');
         }
+        
+        
     }
 }
